refactor: export TagChip from shared index and use consistent imports
Exports TagChip from the shared components index file for consistency
with other shared components. Updates imports in ExpandableRow and
FeatureDetailsGrid to use the cleaner import pattern.
