+++ src/components/shared/index.ts
@@ -6,3 +6,4 @@ export { LoadingState } from './LoadingState';
 export { EmptyState } from './EmptyState';
 export { ErrorState } from './ErrorState';
 export { ChartTooltip, ChartTooltipText } from './ChartTooltip';
+export { TagChip } from './TagChip';
